requirements.txt
# requirements.txt
# Dependências essenciais do projeto

# Framework web
fastapi==0.115.6
uvicorn==0.34.0

# Validação de dados
pydantic==2.10.0

# Configuração e ambiente
python-dotenv==1.0.1

# Formulários (necessário para Form() do FastAPI)
python-multipart==0.0.20

# Type hints
typing_extensions==4.12.2

# Kafka
kafka-python>=2.0.0

# Servidor WSGI para produção
gunicorn==23.0.0

# Nota: As seguintes dependências são instaladas automaticamente como dependências transitivas:
# - starlette (framework base do FastAPI)
# - anyio, sniffio (async)
# - httptools, websockets, h11, httpcore, httpx (HTTP)
# - click, rich, typer, shellingham (CLI tools - usados pelo fastapi-cli)
# - certifi, idna (SSL/IDN)
# - annotated-types, pydantic_core (Pydantic)
# - watchfiles (file watching para reload)
# - MarkupSafe (segurança de templates)
